cleaned up integration test
