+++ integration-tests/jest.config.js
@@ -1,4 +1,4 @@
 module.exports = {
   preset: 'jest-puppeteer',
-  testRegex: 'preview-dos\\.integration-test\\.js$'
+  testRegex: '\\.integration-test\\.js$'
 }
