--- integration-tests/jest/preview-dos.integration-test.js
@@ -1,5 +1,0 @@
-const { registerTest } = require('./jest-utils')
-const { testPreviewDos } = require('../tests/preview-dos')
-
-
-registerTest(testPreviewDos)
