+++ integration-tests/package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "test": "jest --detectOpenHandles",
+    "test": "jest",
     "start": "node src/server.js",
     "start-dev": "nodemon"
   },
